refactor: Fix spotbugs issues

+++ src/main/java/io/github/sudoitir/dddcqrstoolkit/valueobject/Version.java
@@ -3,7 +3,7 @@ package io.github.sudoitir.dddcqrstoolkit.valueobject;
 import java.io.Serial;
 import java.io.Serializable;
 
-public record Version(long version) implements Comparable<Version>, Serializable {
+public record Version(int version) implements Comparable<Version>, Serializable {
 
     @Serial
     private static final long serialVersionUID = 5412681989720822033L;
@@ -20,12 +20,12 @@ public record Version(long version) implements Comparable<Version>, Serializable
     }
 
     public static Version zero() {
-        return new Version(0L);
+        return new Version(0);
     }
 
     @Override
     public int compareTo(final Version val) {
-        return Long.compare(this.version, val.version);
+        return Integer.compare(this.version, val.version);
     }
 
 }
